mjs -> js. Then add .cjs for node shims
