+++ lib/capability-tool.cjs
@@ -1,0 +1,2 @@
+const _require = require('esm')(module) // eslint-disable-line
+module.exports = _require('./capability-tool.mjs').default;
