+++ lib/client.cjs
@@ -1,0 +1,3 @@
+const _require = require('esm')(module) // eslint-disable-line
+
+module.exports = _require('./client.js').default;
