+++ lib/deprecations.js
