+++ lib/http-client.js
@@ -173,4 +173,4 @@ export default class {
       requestHeaders,
     );
   }
-};
+}
