+++ lib/logging.js
@@ -7,7 +7,7 @@ function inspectObject(obj) {
   return obj ? JSON.stringify(obj) : null;
 }
 
-export const logRequestError = function(error) {
+export const logRequestError = function logRequestError(error) {
   if (!errorLogger.enabled) { return; }
   if (error.response) {
     errorLogger(`Error: ${error.response.status}`);
@@ -24,23 +24,23 @@ export const logRequestError = function(error) {
   }
 };
 
-export const logRequestInfo = function(action, url, headers) {
+export const logRequestInfo = function logRequestInfo(action, url, headers) {
   if (!infoLogger.enabled) { return; }
   if (url) { infoLogger(`${action.toUpperCase()} ${url.toString()}`); }
   infoLogger(`Headers: ${inspectObject(headers)}`);
 };
 
-export const logResponseInfo = function(response) {
+export const logResponseInfo = function logResponseInfo(response) {
   if (!infoLogger.enabled) { return; }
   infoLogger(`Response: ${response.status}`);
-  if(response.requestResponse) {
+  if (response.requestResponse) {
     infoLogger(inspectObject(response.requestResponse));
   } else {
     infoLogger(inspectObject(response));
   }
 };
 
-export const logError = function(error) {
+export const logError = function logError(error) {
   if (!errorLogger.enabled) { return; }
   errorLogger(error);
 };
