+++ lib/reference-resolver.js
@@ -144,4 +144,4 @@ export default class {
     }
     return entry.resource;
   }
-};
+}
