+++ lib/smart.js
@@ -1,7 +1,9 @@
-import universalUrl from 'universal-url'; universalUrl.shim();
+import universalUrl from 'universal-url';
+import { logError } from './logging';
+
+universalUrl.shim();
 
 const smartOauthUrl = 'http://fhir-registry.smarthealthit.org/StructureDefinition/oauth-uris';
-import { logError } from './logging';
 
 /**
  * Get the SMART OAuth URIs from a Capability Statement
@@ -20,7 +22,7 @@ import { logError } from './logging';
  * @return {Object} contains the following SMART URIs: authorizeUrl, tokenUrl,
  *   registerUrl, manageUrl
  */
-export const authFromCapability = function(capabilityStatement) {
+export const authFromCapability = function authFromCapability(capabilityStatement) {
   const authMetadata = {};
 
   try {
@@ -50,4 +52,4 @@ export const authFromCapability = function(capabilityStatement) {
     logError(error);
     return authMetadata;
   }
-}
+};
