+++ lib/utils.js
@@ -1,5 +1,5 @@
-import { fhirReferenceRegEx } from './fhir-regular-expressions';
 import queryString from 'query-string';
+import { fhirReferenceRegEx } from './fhir-regular-expressions';
 
 /**
  * Split a FHIR reference into its baseUrl (if present), type, and id
@@ -19,7 +19,7 @@ import queryString from 'query-string';
  * @return {Object} Contains baseUrl, resourceType, and id fields. baseUrl will be
  * undefined for relative references
  */
-export const splitReference = function(reference) {
+export const splitReference = function splitReference(reference) {
   if (!reference.match(fhirReferenceRegEx)) {
     throw new Error(`${reference} is not a recognized FHIR reference`);
   }
@@ -40,7 +40,7 @@ export const splitReference = function(reference) {
   };
 };
 
-export const createQueryString = function(queryParams) {
+export const createQueryString = function createQueryString(queryParams) {
   if (queryParams instanceof Object && Object.keys(queryParams).length > 0) {
     return queryString.stringify(queryParams);
   }
